@@ -9,15 +9,16 @@
   },
   "license": "agpl",
   "private": true,
+  "type": "module",
   "scripts": {
-    "build": "webpack --node-env production --progress",
-    "dev": "webpack --node-env development --progress",
-    "watch": "webpack --node-env development --progress --watch",
-    "serve": "webpack --node-env development serve --progress",
-    "lint": "eslint --ext .js,.vue src",
-    "lint:fix": "eslint --ext .js,.vue src --fix",
-    "stylelint": "stylelint css/*.css css/*.scss src/**/*.scss src/**/*.vue",
-    "stylelint:fix": "stylelint css/*.css css/*.scss src/**/*.scss src/**/*.vue --fix",
+    "dev": "vite",
+    "build": "run-p type-check \"build-only {@}\" --",
+    "postinstall": "npm run build",
+    "preview": "vite preview",
+    "build-only": "vite build",
+    "type-check": "vue-tsc --build --force",
+    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix --ignore-path .gitignore",
+    "format": "prettier --write src/",
     "docker:build": "./bin/build-docker.sh",
     "docker:run": "./bin/run-docker.sh",
     "docker:stop": "./bin/stop-docker.sh",
@@ -26,40 +27,42 @@
   },
   "dependencies": {
     "@churchtools/churchtools-client": "^1.3.11",
-    "@nextcloud/axios": "^2.4.0",
-    "@nextcloud/calendar-js": "^6.1.0",
-    "@nextcloud/dialogs": "^3.1.2",
-    "@nextcloud/initial-state": "^2.1.0",
-    "@nextcloud/l10n": "^2.2.0",
-    "@nextcloud/router": "^2.2.1",
-    "@nextcloud/vue": "^7.12.7",
-    "@nextcloud/vue-richtext": "^2.1.0-beta.5",
+    "@nextcloud/axios": "^2.5.1",
+    "@nextcloud/initial-state": "^2.2.0",
+    "@nextcloud/l10n": "^3.1.0",
+    "@nextcloud/router": "^3.0.1",
+    "@nextcloud/vue": "^9.0.0-alpha.5",
     "pinia": "^2.1.7",
     "v-click-outside": "^3.2.0",
-    "vue": "^2.7.16",
+    "vue": "^3.4.29",
     "vue-material-design-icons": "^5.3.0"
   },
-  "browserslist": [
-    "extends @nextcloud/browserslist-config"
-  ],
   "engines": {
-    "node": "^16.0.0",
-    "npm": "^7.0.0 || ^8.0.0"
+    "node": "^20.0.0",
+    "npm": "^10.8.2"
   },
   "devDependencies": {
-    "@nextcloud/babel-config": "^1.0.0",
-    "@nextcloud/browserslist-config": "^2.2.0",
-    "@nextcloud/eslint-config": "^8.0.0",
-    "@nextcloud/stylelint-config": "^2.1.2",
-    "@nextcloud/webpack-vue-config": "^5.2.1",
-    "autoprefixer": "^9.8.8",
-    "eslint-webpack-plugin": "^4.1.0",
-    "postcss": "^8.4.38",
-    "postcss-loader": "^8.1.1",
-    "postcss-preset-env": "^9.5.9",
-    "stylelint-webpack-plugin": "^4.1.1",
-    "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.2.17",
-    "webpack": "^5.91.0",
-    "webpack-cli": "^5.1.4"
+    "@rushstack/eslint-patch": "^1.8.0",
+    "@tsconfig/node20": "^20.1.4",
+    "@types/node": "^20.14.5",
+    "@vitejs/plugin-vue": "^5.0.5",
+    "@vue/eslint-config-prettier": "^9.0.0",
+    "@vue/eslint-config-typescript": "^13.0.0",
+    "@vue/tsconfig": "^0.5.1",
+    "autoprefixer": "^10.4.20",
+    "cssnano": "^7.0.6",
+    "eslint": "^8.57.0",
+    "eslint-plugin-vue": "^9.23.0",
+    "npm-run-all2": "^6.2.0",
+    "postcss": "^8.4.47",
+    "prettier": "^3.2.5",
+    "simple-xml-to-json": "^1.2.3",
+    "tailwindcss": "^3.4.13",
+    "typescript": "~5.4.0",
+    "vite": "^5.3.1",
+    "vite-plugin-node-polyfills": "^0.22.0",
+    "vite-plugin-vue-devtools": "^7.3.1",
+    "vue-tsc": "^2.0.21",
+    "xml2js": "^0.6.2"
   }
 }
