@@ -1,6 +1,7 @@
 <?php
 $appId = OCA\churchtoolsintegration\AppInfo\Application::APP_ID;
-\OCP\Util::addScript($appId, $appId . '-adminSettings');
+\OCP\Util::addScript($appId, '../dist/' . $appId . '-adminSettings');
+\OCP\Util::addStyle($appId, '../dist/' . $appId . '-adminSettings');
 ?>
 
 <div id="ct_prefs"></div>
